+++ src/main/java/ma/vi/datalines/html/HtmlTableLineReader.java
@@ -11,9 +11,9 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Iterator;
-import java.util.List;
+import java.util.LinkedHashMap;
+import java.util.Map;
 
 /**
  * A line reader which can read data from an HTML table.
@@ -22,8 +22,8 @@ import java.util.List;
  */
 public class HtmlTableLineReader extends AbstractLineReader {
   @Override
-  public boolean supports(File inputFile, String name, Format format) {
-    try (BufferedReader in = new BufferedReader(new FileReader(inputFile))) {
+  public boolean supports(File file, String name, Format format) {
+    try (BufferedReader in = new BufferedReader(new FileReader(file))) {
       String line;
       while ((line = in.readLine()) != null && line.trim().length() == 0);
       if (line != null) {
@@ -50,23 +50,24 @@ public class HtmlTableLineReader extends AbstractLineReader {
   }
 
   @Override
-  protected List<Object> nextLine(boolean convertToColumnType) {
+  protected Map<String, Object> nextLine(boolean convertToColumnType) {
     if (rows.hasNext()) {
-      List<Object> row = new ArrayList<>();
+      int i = 1;
+      Map<String, Object> row = new LinkedHashMap<>();
       Elements cells = rows.next().children();
       for (Element cell : cells) {
         Elements children = cell.children();
         if (children.size() > 0) {
           Element first = children.first();
           if (first.hasAttr("href")) {
-            row.add(first.absUrl("href"));
+            row.put(String.valueOf(i++), first.absUrl("href"));
           } else {
-            row.add(cell.text());
+            row.put(String.valueOf(i++), cell.text());
           }
         } else if (cell.childNodes().size() > 0) {
-          row.add(cell.text());
+          row.put(String.valueOf(i++), cell.text());
         } else {
-          row.add(null);
+          row.put(String.valueOf(i++), null);
         }
       }
       return row;